requirements.txt
pyOpenSSL==22.0.0
zeep==4.1.0
